split up download service more
